Add simple search functionality for notes
I want to demonstrate how to write tests for class methods and scopes,
but the sample code so far doesn't lend itself to that. Add a scope for
searching notes for a keyword, and a form for searching a project's
notes. Use the existing notes#index action to make the scope accessible.